pyproject.toml
[tool.poetry]
name = "wallet"
version = "0.1.0"
description = ""
authors = ["dimitrionian-II <dmitry.bezlutzky@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
django = "^5.0.6"
djangorestframework = "3.15.2"
markdown = "3.6"
django-filter = "24.2"
pytest-django = "4.8.0"
factory-boy = "3.3.0"
drf-spectacular = {extras = ["sidecar"], version = "^0.27.2"}
django-extensions = "3.2.3"
django-cors-headers = "4.4.0"
pre-commit = "^3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
